MA-22138 remove forcedDensity property in display_settings.xml
due to the default launcher3 SecondaryDisplayLauncher navigation bar buttons are
transparent, "forcedDensity = 160" in the secondary displays will make navigation
bar buttons all be moved to the left, which will cause confusion to use.

later will check how to make the navigation bar buttons become visible, but before
the fix, put home button in the middle and back button in the left will make the
usage be closer to user habits.

Change-Id: I5e78942e84fec703ab0c8169754a8111ada8c231
Signed-off-by: Dandan Sun <dandan.sun@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/24025
Reviewed-by: Bo Zhang <bo.zhang@nxp.com>