MA-22198 Add soong build configs for hwcomposer3 to separate some limitations
The Overlay feature and DPU(2d-composition) can be enabled on imx95 at the same
time. The Overlay of DPU is different from the one in DCSS(imx8mq). Need to
separate the limitation of them for different boards.
There is a limitation about the nv12 full range video is only valid on
imx8mm/imx8mp/imx8ulp, but not valid on imx8q and imx95 when do 2d-composition.
So also need to separate the limitation of 2d-composition between g2d-viv.so and
g2d-dpu.so.

Change-Id: I49396beb474c7b503880229f87a7430af6c712e5
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/24194
Reviewed-by: Jessie Hao <juan.hao@nxp.com>