MA-22361-5 Update cma heap range
For widevine L1, because it supports 4K video and
the VPU driver also needs to allocate secure memory, 256M
is not enough. It has been increased to 335M and it is ok
to play 4K secure video. At the same time, this memory is
provided for confirmationUI and widevine to be used together.

The address range of the updated CMA is from 0xB50M ~0xFF0M,
0xA00M ~0xB50M is as secure memory(secure heap and vpu secure boot memroy),
so if the CMA is expanded, the CMA allocation will fail.

Change-Id: Ib3ca26ecbb3b925f71cd30134e6501ed9445efdc
Signed-off-by: He Zhai <zhai.he@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/24771
Reviewed-by: Bo Zhang <bo.zhang@nxp.com>